OAuth.sh uses md5 command if no md5sum
The script uses the GNU `md5sum` command to generate a nonce.
While this is standard on most linux distributions, BSD variants come with
the equivalent `md5` instead (including Mac OS X). This is a small change
to allow the script to work out-of-the-box on most BSD variants, instead of
having to install `md5sum`.